requirements.txt
paho-mqtt==1.6.1
pymongo==4.3.3
mysql-connector-python==8.0.33
neo4j==5.2.0
keyboard==0.13.5